% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreparationSPoTApi.r
\name{PreparationSPoTApi_add_preparations}
\alias{PreparationSPoTApi_add_preparations}
\title{Add new preparation object}
\usage{
PreparationSPoTApi_add_preparations(source, ...)
}
\arguments{
\item{source}{}
}
\description{
Add new preparation object
}
