% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreparationSPoTApi.r
\name{PreparationSPoTApi_delete_preparation}
\alias{PreparationSPoTApi_delete_preparation}
\title{Delete a preparation object}
\usage{
PreparationSPoTApi_delete_preparation(id, ...)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}
}
\description{
Delete a preparation object
}
