% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreparationSPoTApi.r
\name{PreparationSPoTApi_delete_preparation_group}
\alias{PreparationSPoTApi_delete_preparation_group}
\title{Delete a preparation group with all related preparation objects}
\usage{
PreparationSPoTApi_delete_preparation_group(id, ...)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}
}
\description{
Delete a preparation group with all related preparation objects
}
