% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreparationSPoTApi.r
\name{PreparationSPoTApi_update_preparation}
\alias{PreparationSPoTApi_update_preparation}
\title{Update a preparation object}
\usage{
PreparationSPoTApi_update_preparation(id, body, ...)
}
\arguments{
\item{id}{Unique identifier (accession) of the object.}

\item{body}{Metadata in the form of \code{{key: value, key2: value2, ...}}}
}
\description{
Update a preparation object
}
